requirements.txt
flask==3.0.3
openai==1.37.0
python-dotenv==1.0.1
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.0
flask-limiter==3.8.0
